Hatay State
Hatay State (Turkish: Hatay Devleti, French: État du Hatay, Arabic: دولة خطاي Dawlat Khaṭāy), also known informally as the Republic of Hatay, was a transitional political entity that existed from 7 September 1938 to 29 June 1939, being located in the territory of the Sanjak of Alexandretta of the French Mandate of Syria. The state was transformed de jure into the Hatay Province of Turkey on 7 July 1939, de facto joining the country on 23 July 1939.