core shenanigans
A) crimea is not a core of ukraine in vanilla hoi4.
B) it was not owned by the ukssr until cornlord transferred it in 1954.
C) https://media.discordapp.net/attachments/639114087166640148/1114989471897759784/pic_U_K_Ukrainian_SSR_Map.jpg?width=1082&height=670
also, fixed ukrainian cores in poland to be the ones pre-transfer and fixed generalgouvernment borders in galicia to adjust for such